requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
mysqlclient==2.2.4
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
